Revert "Add Firefox user-agent extension."
This reverts commit c76e89767e576834aeda6b92a26e5050a671b2e3.
The extension is nice but it seems sometimes it cannot persist the settings.
Instead, on Linux, one can use the `general.useragent.override` in `about:config` for the same results.
I am considering adding it as a config for NixOS, but for the moment I will leave it as a manual step.

An example: Mozilla/5.0 (Wayland; Linux) AppleWebKit/537.36 (KHTML, like Gecko) Chromium/88.0.4324.150 Safari/537.36 Firefox/138.0